consolidated db functions and minor things
