+++ app.js
@@ -2,12 +2,12 @@ require('dotenv').config();
 const express = require('express');
 const cron = require('node-cron');
 const router = require('./routes/router');
-const routerFunctions = require('./routes/router.helper.functions');
 const functions = require('./utils/functions');
 const auth = require('./auth/auth');
 const authFunctions = require('./auth/auth.helper.functions');
 const cookieParser = require('cookie-parser');
-const email = require('./email/email')
+const email = require('./email/email');
+const db = require('./utils/db');
 
 const app = express();
 const port = process.env.PORT || 8080;
@@ -16,19 +16,9 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 app.use(cookieParser());
 
-app.use('/api', router);
 app.use('/auth', auth)
+//add in routes for serving static files
 app.use(express.static('public'));
-
-//build in logic for showing user front-end based on cookie
-
-// app.use((req, res, next) => {
-//     // Access the refresh token from the cookie
-//     const refreshToken = req.cookies.refresh_token;
-//     console.log("Refresh Token:", refreshToken);
-//     // "next()" tells express to jump to the next middleware function in the app, load before routes
-//     next();
-// })
 
 app.listen(port, () => {
     console.log(`Server is running on port ${port}`);
@@ -36,7 +26,7 @@ app.listen(port, () => {
 
 async function run() {
 
-    const users = await routerFunctions.getUsers();
+    const users = await db.getUsers();
 
     await authFunctions.refreshAccessTokens(users);
 
@@ -56,7 +46,7 @@ async function run() {
     }
 }
 
-// run().catch(console.error);
+run().catch(console.error);
 
 // schedule the run() function to occur every Friday at 9am, Central Standard Time
 const scheduledRun = cron.schedule('* 9 * * Fri', () => {
@@ -67,6 +57,6 @@ const scheduledRun = cron.schedule('* 9 * * Fri', () => {
     timezone: "US/Central"
 });
 
-scheduledRun.start();
+// scheduledRun.start();
